Extracts prosecutor selection component
