+++ apps/judicial-system/web/src/routes/Prosecutor/SharedComponents/ProsecutorSection/ProsecutorSection.strings.ts
@@ -7,18 +7,6 @@ export const strings = defineMessages({
     description:
       'Notaður sem titill fyrir "ákærandi" hlutann á óskir um fyrirtöku skrefi í öllum málstegundum.',
   },
-  placeholder: {
-    id: 'judicial.system.core:select_prosecutor.placeholder',
-    defaultMessage: 'Veldu sækjanda',
-    description:
-      'Notaður sem skýritexti í "Veldu sækjanda" textaboxi á óskir um fyrirtöku skrefi í öllum málstegundum.',
-  },
-  label: {
-    id: 'judicial.system.core:select_prosecutor.label',
-    defaultMessage: 'Veldu sækjanda',
-    description:
-      'Notaður sem titill í "Veldu sækjanda" textaboxi á óskir um fyrirtöku skrefi í öllum málstegundum.',
-  },
   tooltip: {
     id: 'judicial.system.core:select_prosecutor.tooltip',
     defaultMessage:
