+++ apps/judicial-system/web/src/routes/Prosecutor/SharedComponents/ProsecutorSection/ProsecutorSelection.strings.ts
@@ -1,0 +1,16 @@
+import { defineMessages } from 'react-intl'
+
+export const strings = defineMessages({
+  placeholder: {
+    id: 'judicial.system.core:prosecutor_selection.placeholder',
+    defaultMessage: 'Veldu sækjanda',
+    description:
+      'Notaður sem skýritexti í "Veldu sækjanda" textaboxi á óskir um fyrirtöku skrefi í öllum málstegundum.',
+  },
+  label: {
+    id: 'judicial.system.core:prosecutor_selection.label',
+    defaultMessage: 'Veldu sækjanda',
+    description:
+      'Notaður sem titill í "Veldu sækjanda" textaboxi á óskir um fyrirtöku skrefi í öllum málstegundum.',
+  },
+})
